+++ examples/04_glm_first_level_models/plot_spm_multimodal_faces.py
@@ -25,7 +25,7 @@ print(__doc__)
 
 #########################################################################
 # Fetch the SPM multimodal_faces data.
-from nilearn.datasets.func import fetch_spm_multimodal_fmri
+from nilearn.datasets import fetch_spm_multimodal_fmri
 subject_data = fetch_spm_multimodal_fmri()
 
 #########################################################################